feat: 선호도조사, update: 마이페이지 누락된 부분 추가
