+++ package.json
@@ -3,6 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.9",
+    "@mui/material": "^5.15.9",
     "@reduxjs/toolkit": "^2.1.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
@@ -17,6 +20,7 @@
     "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
     "react-select": "^5.8.0",
+    "redux": "^5.0.1",
     "redux-persist": "^6.0.0",
     "styled-component": "^2.8.0",
     "styled-components": "^6.1.8",
