+++ src/App.js
@@ -1,14 +1,31 @@
 import { Route, Routes } from "react-router-dom";
 import MainPageComponent from "./Components/MainPage/MainPageComponent";
-import MyPageComponent from "./Pages/MyPage";
-import EditProfile from "./Components/myPage/EditProfile";
-import ManageShippingRefund from "./Components/myPage/ManageShippingRefund";
-import NotificationSettings from "./Components/myPage/NotificationSettings";
-import PasswordChange from "./Components/myPage/PasswordChange";
-import MemberWithdrawal from "./Components/myPage/MemberWithdrawal";
-import ManagePurchase from "./Components/myPage/ManagePurchase";
+// 마이페이지(고객님)
+import MyPageCustomerComponent from "./Pages/MyPageCustomer";
+import EditProfile from "./Components/myPage/myPage(Customer)/EditProfile";
+import ManageShippingRefund from "./Components/myPage/myPage(Customer)/ManageShippingRefund";
+import ManagePurchase from "./Components/myPage/myPage(Customer)/ManagePurchase";
+import CustomizedInformation from "./Components/myPage/myPage(Customer)/CustomizedInformation";
+import NotificationSettings from "./Components/myPage/myPage(Customer)/NotificationSettings";
+import PasswordChange from "./Components/myPage/myPage(Customer)/PasswordChange";
+import MemberWithdrawal from "./Components/myPage/myPage(Customer)/MemberWithdrawal";
+import MyReview from "./Components/myPage/myPage(Customer)/MyReview";
+// 마이페이지(사장님)
+import MyPageSellerComponent from "./Pages/MyPageSeller";
+import SellerProfile from "./Components/myPage/myPage(Seller)/SellerProfile";
+import EditProfileSeller from "./Components/myPage/myPage(Seller)/EditProfileSeller";
+import ManagePurchaseSeller from "./Components/myPage/myPage(Seller)/ManagePurchaseSeller";
+import AlarmSeller from "./Components/myPage/myPage(Seller)/NotificationSettingsSeller";
+import PasswordChangeSeller from "./Components/myPage/myPage(Seller)/PasswordChangeSeller";
+import MemberWithdrawalSeller from "./Components/myPage/myPage(Seller)/MemberWithdrawalSeller";
+import MyReviewSeller from "./Components/myPage/myPage(Seller)/MyReviewSeller";
+import ProfitSeller from "./Components/myPage/myPage(Seller)/ProfitSeller";
+// 선호도 조사
+import Preference from "./Pages/Preference";
+// 포스트
 import Post from "./Pages/Post";
 import Seller from "./Pages/Seller";
+// 로그인
 import Login from "./Pages/Login";
 import Register from "./Pages/Register";
 import RegisterForm from "./Components/Register/RegisterForm";
@@ -30,9 +47,73 @@ function App() {
         <div className="App">
           <Routes>
             <Route path="/" element={<MainPageComponent />} />
-            <Route path="/*" element={<Post />} />
+            <Route path="/preference" element={<Preference />} />
+            {/* <Route path="/*" element={<Post />} /> */}
+            {/* Seller 부분 */}
+            <Route path="/mypageSeller" element={<MyPageSellerComponent />}>
+              <Route
+                path="/mypageSeller/EditProfile"
+                element={<EditProfileSeller />}
+              />
+              <Route
+                path="/mypageSeller/SellerProfile"
+                element={<SellerProfile />}
+              />
+              <Route
+                path="/mypageSeller/ManagePurchase"
+                element={<ManagePurchaseSeller />}
+              />
+              <Route
+                path="/mypageSeller/NotificationSettings"
+                element={<AlarmSeller />}
+              />
+              <Route
+                path="/mypageSeller/PasswordChange"
+                element={<PasswordChangeSeller />}
+              />
+              <Route
+                path="/mypageSeller/MemberWithdrawal"
+                element={<MemberWithdrawalSeller />}
+              />
+              <Route
+                path="/mypageSeller/MyReview"
+                element={<MyReviewSeller />}
+              />
+              <Route
+                path="/mypageSeller/ProfitSeller"
+                element={<ProfitSeller />}
+              />
+            </Route>
+            <Route path="/mypage" element={<MyPageCustomerComponent />}>
+              <Route path="/mypage/EditProfile" element={<EditProfile />} />
+              <Route
+                path="/mypage/ManagePurchase"
+                element={<ManagePurchase />}
+              />
+              <Route
+                path="/mypage/ManageShippingRefund"
+                element={<ManageShippingRefund />}
+              />
+              <Route
+                path="/mypage/CustomizedInformation"
+                element={<CustomizedInformation />}
+              />
+              <Route
+                path="/mypage/NotificationSettings"
+                element={<NotificationSettings />}
+              />
+              <Route
+                path="/mypage/PasswordChange"
+                element={<PasswordChange />}
+              />
+              <Route
+                path="/mypage/MemberWithdrawal"
+                element={<MemberWithdrawal />}
+              />
+              <Route path="/mypage/MyReview" element={<MyReview />} />
+            </Route>
             <Route path="/seller" element={<Seller />} />
-            <Route path="/mypage" element={<MyPageComponent />}>
+            <Route path="/mypage" element={<MyPageCustomerComponent />}>
               <Route path="/mypage/EditProfile" element={<EditProfile />} />
               <Route
                 path="/mypage/ManagePurchase"
