+++ src/store.js
@@ -3,9 +3,11 @@ import { combineReducers } from "redux";
 import { persistReducer, persistStore } from "redux-persist";
 import storageSession from "redux-persist/lib/storage/session";
 import loginReducer from "./redux/loginSlice";
+import preferenceSlice from "./redux/preferenceSlice";
 
 const rootReducer = combineReducers({
   login: loginReducer,
+  preference: preferenceSlice,
 });
 
 const persistConfig = {
